# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  pkg_api_handlers.APIKeyCreatedResponse:
    description: API key creation response (includes secret, shown only once)
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      expires_at:
        example: "2024-04-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      key:
        example: dg_live_abc123xyz789...
        type: string
      last_used_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      name:
        example: CI/CD Key
        type: string
      prefix:
        example: dg_live_abc123
        type: string
      project_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      scope:
        example: user
        type: string
    type: object
  pkg_api_handlers.APIKeyResponse:
    description: API key information (secret only shown on creation)
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      expires_at:
        example: "2024-04-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_used_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      name:
        example: CI/CD Key
        type: string
      prefix:
        example: dg_live_abc123
        type: string
      project_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      scope:
        example: user
        type: string
    type: object
...
  pkg_api_handlers.CreateAPIKeyRequest:
    description: Create API key request
    properties:
      expires_in:
        example: 90d
        type: string
      name:
        example: CI/CD Key
        type: string
    required:
    - name
    type: object
...
  pkg_api_handlers.DashboardOverviewResponse:
    description: Dashboard overview response
    properties:
      projects:
        items:
          $ref: '#/definitions/pkg_api_handlers.DashboardProjectResponse'
        type: array
      recent_runs:
        items:
          $ref: '#/definitions/pkg_api_handlers.DashboardRunResponse'
        type: array
    type: object
  pkg_api_handlers.DashboardProjectResponse:
    description: Dashboard project with inline stats
    properties:
      avg_duration_ms:
        type: integer
      chart:
        items:
          $ref: '#/definitions/pkg_api_handlers.RunDashboardChartPointResponse'
        type: array
      created_at:
        type: string
      failed_runs_7d:
        type: integer
      id:
        type: string
      latest_run:
        $ref: '#/definitions/pkg_api_handlers.DashboardRunResponse'
      member_count:
        type: integer
      name:
        type: string
      repo_url:
        type: string
      slug:
        type: string
      success_runs_7d:
        type: integer
      top_branch:
        type: string
      total_runs_7d:
        type: integer
      updated_at:
        type: string
      visibility:
        type: string
    type: object
  pkg_api_handlers.DashboardRunResponse:
    description: Dashboard run entry
    properties:
      commit_author_name:
        type: string
      commit_sha:
        type: string
      created_at:
        type: string
      duration_ms:
        type: integer
      id:
        type: string
      project_id:
        type: string
      project_name:
        type: string
      status:
        type: string
      trigger_ref:
        type: string
      triggered_by_user:
        $ref: '#/definitions/pkg_api_handlers.UserResponse'
      workflow_name:
        type: string
    type: object
...
  pkg_api_handlers.ErrorResponse:
    description: Error response returned by the API
    properties:
      details: {}
      error:
        example: bad_request
        type: string
      message:
        example: Invalid request body
        type: string
    type: object
...
  pkg_api_handlers.RunDashboardChartPointResponse:
    description: Daily chart point for project runs
    properties:
      date:
        example: "2026-02-11"
        type: string
      duration_ms:
        example: 95000
        type: integer
      failed:
        example: 2
        type: integer
      success:
        example: 12
        type: integer
    type: object
...
  pkg_api_handlers.UserResponse:
    description: User information
    properties:
      avatar_url:
        example: https://avatars.githubusercontent.com/u/1234567
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
      provider:
        example: github
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@dagryn.dev
    name: Dagryn Support
    url: https://github.com/mujhtech/dagryn
  description: Local-first, self-hosted developer workflow orchestrator API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Dagryn API
  version: "1.0"
paths:
...
  /api/v1/dashboard/overview:
    get:
      description: Returns projects with inline stats and recent runs across all user
        projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.DashboardOverviewResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard overview
      tags:
      - dashboard
...
  /api/v1/projects/{projectId}/api-keys:
    get:
      description: Returns all API keys for a project
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pkg_api_handlers.APIKeyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: List project API keys
      tags:
      - projects
      - api-keys
    post:
      consumes:
      - application/json
      description: Creates a new API key for a project
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: projectId
        required: true
        type: string
      - description: Create API key request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_api_handlers.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_api_handlers.APIKeyCreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: Create project API key
      tags:
      - projects
      - api-keys
  /api/v1/projects/{projectId}/api-keys/{keyID}:
    delete:
      description: Revokes an API key for a project
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: projectId
        required: true
        type: string
      - description: API Key ID
        format: uuid
        in: path
        name: keyID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: Revoke project API key
      tags:
      - projects
      - api-keys
...
securityDefinitions:
  APIKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"